Begin implementation of Confirmation button
Starting work on there being a confirmation button when irreversible
requests are made eg when user wants to delete an event.

Main reasoning is to add a step before request is sent where user answers
if they are sure they want to proceed. If they do, we proceed else,
we dont send request.