Fix multiple confirmation boxes appearing
Just added a counter variable. If its not zero it wont spawn another
confirmation box if delete button is clicked.

I will likely add similar to places where i want something clicked once.

Also fixed bug in new_save_with_overlap. If new event has same start time
with another event an error will now be raised as it is ambiguous the way
to shift events (tbf this is slightly a limitation of the way i implemented
it but it can be justifiable).